API Reference

TGLFNN

TGLFNN.compare_two_input_tglfs — Function

compare_two_input_tglfs(itp_1::TGLFNN.InputTGLF, itp_2::TGLFNN.InputTGLF)

Compares two input_tglfs, prints the difference and stores the difference in a new InputTGLF

TGLFNN.run_qlgyro — Function

input_qlgyro(input_qlgyro::InputQLGYRO, input_cgyro::InputCGYRO)

Run QLGYRO starting from a InputQLGYRO and InputCGYRO

Returns a flux_solution structure

run_qlgyro(input_qlgyros::Vector{InputQLGYRO}, input_cgyros::Vector{InputCGYRO})

Run QLGYRO starting from a vectors of InputQLGYRO and InputCGYRO

NOTE: Each run is done sequentially, one after the other

Returns a vector of flux_solution structures

TGLFNN.run_tglf — Function

run_tglf(input_tglf::InputTGLF)

Run TGLF starting from a InputTGLF.

Returns a flux_solution structure

run_tglf(input_tglf::InputTGLF)

Run TGLF starting from a vector of InputTGLFs.

NOTE: Each run is done asyncronously (ie. in separate parallel processes)

Returns a flux_solution structure

TGLFNN.run_tglfnn — Function

run_tglfnn(input_tglf::InputTGLF; model_filename::String, uncertain::Bool=false, warn_nn_train_bounds::Bool)

Run TGLFNN starting from a InputTGLF, using a specific model_filename.

If the model is an ensemble of NNs, then the output can be uncertain (using the Measurements.jl package).

The warnnntrain_bounds checks against the standard deviation of the inputs to warn if evaluation is likely outside of training bounds.

Returns a flux_solution structure

run_tglfnn(input_tglfs::Vector{InputTGLF}; model_filename::String, uncertain::Bool=false, warn_nn_train_bounds::Bool)

Run TGLFNN for multiple InputTGLF, using a specific model_filename.

This is more efficient than running TGLFNN on each individual InputTGLFs.

If the model is an ensemble of NNs, then the output can be uncertain (using the Measurements.jl package).

The warnnntrain_bounds checks against the standard deviation of the inputs to warn if evaluation is likely outside of training bounds.

Returns a vector of flux_solution structures

run_tglfnn(data::Dict; model_filename::String, uncertain::Bool=false, warn_nn_train_bounds::Bool)::Dict

Run TGLFNN from a dictionary, using a specific model_filename.

If the model is an ensemble of NNs, then the output can be uncertain (using the Measurements.jl package).

The warnnntrain_bounds checks against the standard deviation of the inputs to warn if evaluation is likely outside of training bounds.

Returns a dictionary with fluxes